Refactor: isolate multi-node editor presentation
